break down pelias-import into smaller steps in hopes of making retries less painful.
I'm seeing an inexplicable failure between the OSM and WOF import - the
WOF import can't find the pelias_schema which was created.

Interestingly this happens on the planet build (after 9 hours!) but the
Seattle build doesn't hit the error at all.

Also of note is that the WOF import wasn't added until more recently
(234886e), so I'm not sure if it has ever been tested at the planet
scale.