Compress (some of) the artifacts before upload.
Some very anecdotal evidence shows we'll use about 1/2 the size overall.

As well as save a bit of storage cost, the hope is that this should improve
upload and download speeds during deploy enough to offset the cpu cost of
compression.

This is especially helpful for placeholder, valhalla, and elasticsearch.  pbf
and mbtiles don't seem to benefit.

xz is a bit slower (IIRC maybe twice as slow, but not like 10x slower),
but it seems worth it vs. gz.

Some relative file sizes:
```
423M Seattle.elasticsearch.tar
250M Seattle.elasticsearch.tar.gz
204M Seattle.elasticsearch.tar.xz
   0 Seattle.graph.obj
  38 Seattle.graph.obj.gz
  32 Seattle.graph.obj.xz
 10K Seattle.gtfs.tar
  62 Seattle.gtfs.tar.gz
 108 Seattle.gtfs.tar.xz
 30M Seattle.mbtiles
 30M Seattle.mbtiles.gz
 30M Seattle.mbtiles.xz
 39M Seattle.osm.pbf
 39M Seattle.osm.pbf.gz
 39M Seattle.osm.pbf.xz
1019 Seattle.pelias.json
 459 Seattle.pelias.json.gz
 508 Seattle.pelias.json.xz
545M Seattle.placeholder.tar
144M Seattle.placeholder.tar.gz
 82M Seattle.placeholder.tar.xz
219M Seattle.valhalla.tar
128M Seattle.valhalla.tar.gz
 83M Seattle.valhalla.tar.xz
```